# api/openapi.json
{
    "components": {
        "schemas": {
...
            "api.RepositoryCollectionResponse": {
                "properties": {
                    "data": {
                        "description": "Requested Data",
                        "items": {
                            "$ref": "#/components/schemas/api.RepositoryResponse"
                        },
                        "type": "array"
                    },
                    "links": {
                        "$ref": "#/components/schemas/api.Links"
                    },
                    "meta": {
                        "$ref": "#/components/schemas/api.ResponseMetadata"
                    }
                },
                "type": "object"
            },
            "api.RepositoryIntrospectRequest": {
                "properties": {
                    "reset_count": {
                        "description": "Reset the failed introspections count",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
            "api.RepositoryRequest": {
                "properties": {
                    "distribution_arch": {
                        "description": "Architecture to restrict client usage to",
                        "example": "x86_64",
                        "type": "string"
                    },
                    "distribution_versions": {
                        "description": "Versions to restrict client usage to",
                        "example": [
                            "7",
                            "8"
                        ],
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "gpg_key": {
                        "description": "GPG key for repository",
                        "type": "string"
                    },
                    "metadata_verification": {
                        "description": "Verify packages",
                        "type": "boolean"
                    },
                    "name": {
                        "description": "Name of the remote yum repository",
                        "type": "string"
                    },
                    "url": {
                        "description": "URL of the remote yum repository",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "errors.ErrorResponse": {
                "properties": {
                    "errors": {
                        "items": {
                            "$ref": "#/components/schemas/errors.HandlerError"
                        },
                        "type": "array"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "RhIdentity": {
                "in": "header",
                "name": "x-rh-identity",
                "type": "apiKey"
            }
        }
    },
    "info": {
        "contact": {},
        "description": "API of the Content Sources application on [console.redhat.com](https://console.redhat.com)\n",
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0"
        },
        "title": "ContentSourcesBackend",
        "version": "v1.0.0"
    },
    "openapi": "3.0.3",
    "paths": {
...
        "/repositories/": {
            "get": {
                "description": "list repositories",
                "operationId": "listRepositories",
                "parameters": [
                    {
                        "description": "Offset into the list of results to return in the response",
                        "in": "query",
                        "name": "offset",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Limit the number of items returned",
                        "in": "query",
                        "name": "limit",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Comma separated list of architecture to optionally filter-on (e.g. 'x86_64,s390x' would return Repositories with x86_64 or s390x only)",
                        "in": "query",
                        "name": "version",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Comma separated list of versions to optionally filter-on  (e.g. '7,8' would return Repositories with versions 7 or 8 only)",
                        "in": "query",
                        "name": "arch",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filter by compatible arch (e.g. 'x86_64' would return Repositories with the 'x86_64' arch and Repositories where arch is not set)",
                        "in": "query",
                        "name": "available_for_version",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filter by compatible version (e.g. 7 would return Repositories with the version 7 or where version is not set)",
                        "in": "query",
                        "name": "available_for_arch",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Search term for name and url.",
                        "in": "query",
                        "name": "search",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filter repositories by name using an exact match",
                        "in": "query",
                        "name": "name",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filter repositories by name using an exact match",
                        "in": "query",
                        "name": "url",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Sets the sort order of the results",
                        "in": "query",
                        "name": "sort_by",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Comma separated list of statuses to optionally filter on",
                        "in": "query",
                        "name": "status",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.RepositoryCollectionResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/errors.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/errors.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/errors.ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/errors.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "List Repositories",
                "tags": [
                    "repositories"
                ]
            },
            "post": {
                "description": "create a repository",
                "operationId": "createRepository",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/api.RepositoryRequest"
                            }
                        }
                    },
                    "description": "request body",
                    "required": true,
                    "x-originalParamName": "body"
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.RepositoryResponse"
                                }
                            }
                        },
                        "description": "Created",
                        "headers": {
                            "Location": {
                                "description": "resource URL",
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/errors.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/errors.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/errors.ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/errors.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "Create Repository",
                "tags": [
                    "repositories"
                ]
            }
        },
...
        "/repositories/bulk_create/": {
            "post": {
                "description": "bulk create repositories",
                "operationId": "bulkCreateRepositories",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "items": {
                                    "$ref": "#/components/schemas/api.RepositoryRequest"
                                },
                                "type": "array"
                            }
                        }
                    },
                    "description": "request body",
                    "required": true,
                    "x-originalParamName": "body"
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/api.RepositoryResponse"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "Created",
                        "headers": {
                            "Location": {
                                "description": "resource URL",
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/errors.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/errors.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/errors.ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/errors.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "Bulk create repositories",
                "tags": [
                    "repositories"
                ]
            }
        },
...
        "/repositories/{uuid}/introspect/": {
            "post": {
                "operationId": "introspect",
                "parameters": [
                    {
                        "description": "Identifier of the Repository",
                        "in": "path",
                        "name": "uuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "*/*": {
                            "schema": {
                                "$ref": "#/components/schemas/api.RepositoryIntrospectRequest"
                            }
                        }
                    },
                    "description": "request body",
                    "x-originalParamName": "body"
                },
                "responses": {
                    "204": {
                        "description": "Introspection was successfully queued"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/errors.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/errors.ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/errors.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "introspect a repository",
                "tags": [
                    "repositories"
                ]
            }
        },
...
    },
    "servers": [
        {
            "url": "https://api.example.com/api/content-sources/v1.0/"
        }
    ]
}